Rename packages according to google java style guide...
modified:   dojo-dungeon/src/dojo/rooms/rooms/boss/Fehler_Refactoring.java
modified:   dojo-dungeon/src/dojo/rooms/rooms/boss/MyImpRoom.java
modified:   dojo-dungeon/src/dojo/rooms/rooms/riddle/Fehler_Quader.java
modified:   dojo-dungeon/src/dojo/rooms/rooms/riddle/Fehler_Syntax.java
modified:   dojo-dungeon/src/dojo/rooms/rooms/riddle/Implement_MyMonster.java
renamed:    dojo-dungeon/src/studentTasks/fehlerCuboid/Cuboid.java -> dojo-dungeon/src/studenttasks/cuboid/Cuboid.java
renamed:    dojo-dungeon/src/studentTasks/modifyEntities/ModifyEntities.java -> dojo-dungeon/src/studenttasks/modifyentities/ModifyEntities.java
renamed:    dojo-dungeon/src/studentTasks/implementMyMonster/MyMonster.java -> dojo-dungeon/src/studenttasks/mymonster/MyMonster.java
renamed:    dojo-dungeon/src/studentTasks/fehlerRefactoring/ClsToRefactor.java -> dojo-dungeon/src/studenttasks/refactoring/ClsToRefactor.java
renamed:    dojo-dungeon/src/studentTasks/fehlerRefactoring/Notiz_Tutoren.md -> dojo-dungeon/src/studenttasks/refactoring/Notiz_Tutoren.md
renamed:    dojo-dungeon/src/studentTasks/fehlerSyntax/FehlerhafteKlasse.java -> dojo-dungeon/src/studenttasks/syntax/FehlerhafteKlasse.java
